authentication w/ username-password + oauth2 42 + sessions

+++ back-end/src/services/user.service.ts
@@ -1,8 +1,10 @@
-import { Injectable, NotFoundException, UseGuards } from '@nestjs/common';
+import { HttpCode, HttpException, Injectable, NotFoundException, UseGuards } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { request } from 'https';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from '../dtos/in/CreateUser.dto';
 import { User } from '../models/user.entity';
+import { unlinkSync } from 'fs';
 
 @Injectable()
 export class UserService {
@@ -19,12 +21,27 @@ export class UserService {
         user.lastname = payload.lastname;
         user.username = payload.username;
         user.password = payload.password;
+        user.login42 = null;
+        user.isActive = false;
 
         return this.userRepository.save(user);
     }
 
-    getUser(id: string): Promise<User> {
+    create42User(user: User): Promise<User> {
+
+        return this.userRepository.save(user);
+    }
+
+    getUser(user: string): Promise<User> {
+        return this.userRepository.findOne({ username: user });
+    }
+
+    getUserById(id: string): Promise<User> {
         return this.userRepository.findOne(id);
+    }
+
+    getUserBy42Login(user: string): Promise<User> {
+        return this.userRepository.findOne({ login42: user });
     }
 
     async updateUser(body: CreateUserDto, id: string): Promise<User> {
@@ -40,7 +57,21 @@ export class UserService {
     }
 
     async deleteUser(id: string): Promise<void> {
+
+        const user : User = await this.getUserById(id);
+
+        if (!user)
+            return null; // user does not exist
+
+        const path = `/usr/src/app/avatar/${user.login42}.png`;
+
+        try {
+            unlinkSync(path)
+            //file removed
+        } catch (err) {
+            console.error(err)
+        }
+
         await this.userRepository.delete(id);
     }
-
 }
